Complete restructure of PR, see description.
Move to model where there is only SaveTripModal with difference in props passed in.
